requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
chardet==5.2.0
    # via reportlab
cssselect2==0.7.0
    # via svglib
django==5.0.2
    # via
    #   -r requirements.in
    #   django-classy-tags
    #   django-cms
    #   django-filer
    #   django-formtools
    #   django-js-asset
    #   django-polymorphic
    #   django-sekizai
    #   django-treebeard
    #   djangocms-admin-style
    #   easy-thumbnails
django-classy-tags==4.1.0
    # via
    #   -r requirements.in
    #   django-cms
    #   django-sekizai
django-cms==3.11.5
    # via
    #   -r requirements.in
    #   djangocms-attributes-field
    #   djangocms-bootstrap4
    #   djangocms-file
    #   djangocms-googlemap
    #   djangocms-icon
    #   djangocms-link
    #   djangocms-picture
    #   djangocms-snippet
    #   djangocms-style
    #   djangocms-text-ckeditor
    #   djangocms-video
django-filer==3.1.1
    # via
    #   -r requirements.in
    #   djangocms-bootstrap4
    #   djangocms-file
    #   djangocms-googlemap
    #   djangocms-link
    #   djangocms-picture
    #   djangocms-video
django-formtools==2.5.1
    # via django-cms
django-js-asset==2.2.0
    # via django-mptt
django-mptt==0.16.0
    # via -r requirements.in
django-polymorphic==3.1.0
    # via django-filer
django-sekizai==4.1.0
    # via
    #   -r requirements.in
    #   django-cms
django-treebeard==4.7.1
    # via
    #   -r requirements.in
    #   django-cms
    #   djangocms-googlemap
djangocms-admin-style==3.3.0
    # via
    #   -r requirements.in
    #   django-cms
djangocms-attributes-field==3.0.0
    # via
    #   djangocms-bootstrap4
    #   djangocms-file
    #   djangocms-icon
    #   djangocms-link
    #   djangocms-picture
    #   djangocms-style
    #   djangocms-video
djangocms-bootstrap4==3.0.0
    # via -r requirements.in
djangocms-file==3.0.1
    # via -r requirements.in
djangocms-googlemap==2.2.0
    # via -r requirements.in
djangocms-icon==2.1.0
    # via djangocms-bootstrap4
djangocms-link==3.1.1
    # via
    #   -r requirements.in
    #   djangocms-bootstrap4
djangocms-picture==4.1.1
    # via
    #   -r requirements.in
    #   djangocms-bootstrap4
djangocms-snippet==3.1.1
    # via -r requirements.in
djangocms-style==3.1.0
    # via -r requirements.in
djangocms-text-ckeditor==5.1.5
    # via
    #   -r requirements.in
    #   djangocms-bootstrap4
djangocms-video==3.1.0
    # via -r requirements.in
easy-thumbnails[svg]==2.8.5
    # via
    #   -r requirements.in
    #   django-filer
    #   djangocms-picture
html5lib==1.1
    # via
    #   -r requirements.in
    #   djangocms-text-ckeditor
lxml==5.1.0
    # via svglib
packaging==23.2
    # via
    #   django-cms
    #   djangocms-text-ckeditor
pillow==10.2.0
    # via
    #   -r requirements.in
    #   djangocms-text-ckeditor
    #   easy-thumbnails
    #   reportlab
pypng==0.20220715.0
    # via qrcode
pytz==2024.1
    # via -r requirements.in
qrcode==7.4.2
    # via -r requirements.in
reportlab==4.1.0
    # via
    #   easy-thumbnails
    #   svglib
six==1.16.0
    # via
    #   -r requirements.in
    #   html5lib
sqlparse==0.4.4
    # via django
svglib==1.5.1
    # via easy-thumbnails
tinycss2==1.2.1
    # via
    #   cssselect2
    #   svglib
typing-extensions==4.10.0
    # via qrcode
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
